requirements.txt
ansible==2.7.10
asgiref==3.2.3
autopep8==1.5
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
Django==3.0.4
django-ansible==1.0.3
django-crispy-forms==1.9.0
django-filter==2.3.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
ncclient==0.6.7
paramiko==2.7.1
Pillow==7.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pycparser==2.20
PyNaCl==1.3.0
pytz==2019.3
PyYAML==5.3
scp==0.13.2
six==1.14.0
sqlparse==0.3.1
